requirements.txt
absl-py==1.4.0
astor==0.8.1
cycler==0.11.0
gast==0.5.4
gevent==1.4.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
gunicorn==19.9.0
h5py==3.8.0
importlib-metadata==6.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.0.2
numpy==1.17.4
opencv-python==4.1.2.30
protobuf==3.19.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==3.13
six==1.16.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==2.3.0
typing_extensions==4.6.3
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
